requirements.txt
Django==1.8.2
django-extensions==1.5.5
django-guardian==1.3
django-model-utils==2.2
djangorestframework==3.1.3
ipython==3.1.0
six==1.9.0
tornado==4.2
wheel==0.24.0
